refactor: move domain package location
