+++ src/main/java/com/scg/stop/event/dto/response/EventNoticeListElementResponse.java
@@ -1,7 +1,8 @@
-package com.scg.stop.domain.event.dto.response;
+package com.scg.stop.event.dto.response;
 
 
-import com.scg.stop.domain.event.domain.EventNotice;
+import com.scg.stop.event.domain.EventNotice;
+import com.scg.stop.notice.dto.response.NoticeListElementResponse;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,8 +21,8 @@ public class EventNoticeListElementResponse {
     private LocalDateTime updatedAt;
 
     // Entity -> DTO
-    public static com.scg.stop.domain.notice.dto.response.NoticeListElementResponse from(EventNotice eventNotice) {
-        return new com.scg.stop.domain.notice.dto.response.NoticeListElementResponse(
+    public static NoticeListElementResponse from(EventNotice eventNotice) {
+        return new NoticeListElementResponse(
                 eventNotice.getId(),
                 eventNotice.getTitle(),
                 eventNotice.getHitCount(),
