+++ src/main/java/com/scg/stop/notice/domain/Notice.java
@@ -1,5 +1,4 @@
-package com.scg.stop.domain.event.domain;
-
+package com.scg.stop.notice.domain;
 
 import com.scg.stop.domain.file.domain.File;
 import com.scg.stop.global.domain.BaseTimeEntity;
@@ -10,7 +9,7 @@ import lombok.NoArgsConstructor;
 import java.util.ArrayList;
 import java.util.List;
 
-import static jakarta.persistence.CascadeType.REMOVE;
+import static jakarta.persistence.CascadeType.ALL;
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
@@ -18,7 +17,7 @@ import static lombok.AccessLevel.PROTECTED;
 @Entity
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-public class EventNotice extends BaseTimeEntity {
+public class Notice extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
@@ -35,20 +34,23 @@ public class EventNotice extends BaseTimeEntity {
 
     @Column(nullable = false, columnDefinition = "TINYINT(1)")
     private boolean fixed;
-    
-    @OneToMany(fetch = LAZY, mappedBy = "eventNotice", cascade = REMOVE, orphanRemoval = true)
+
+    @OneToMany(fetch = LAZY, mappedBy = "notice", cascade = ALL, orphanRemoval = true)
     private List<File> files = new ArrayList<>();
 
-    private EventNotice(String title, String content, Integer hitCount, boolean fixed, List<File> files) {
+    // private constructor for creating new notice entity
+    private Notice(String title, String content, Integer hitCount, boolean fixed, List<File> files) {
         this.title = title;
         this.content = content;
         this.hitCount = hitCount;
         this.fixed = fixed;
-        files.forEach(file -> file.setEventNotice(this));
+        files.forEach(file -> file.setNotice(this));
+
     }
 
-    public static EventNotice from(String title, String content, boolean fixed, List<File> files) {
-        return new EventNotice(
+    // static method for creating new notice entity
+    public static Notice from(String title, String content, boolean fixed, List<File> files) {
+        return new Notice(
                 title,
                 content,
                 0,
@@ -57,10 +59,15 @@ public class EventNotice extends BaseTimeEntity {
         );
     }
 
-    public void updateEventNotice(String title, String content, boolean fixed) {
+    public void updateNotice(String title, String content, boolean fixed, List<File> files) {
         this.title = title;
         this.content = content;
         this.fixed = fixed;
+        this.files.clear();
+        if (files != null) {
+            this.files.addAll(files);
+            files.forEach(file -> file.setNotice(this));
+        }
     }
 
     public void increaseHitCount() {
