+++ src/main/java/com/scg/stop/notice/dto/response/NoticeResponse.java
@@ -1,19 +1,20 @@
-package com.scg.stop.domain.event.dto.response;
+package com.scg.stop.notice.dto.response;
 
-import com.scg.stop.domain.event.domain.EventNotice;
 import com.scg.stop.domain.file.domain.File;
 import com.scg.stop.domain.file.dto.response.FileResponse;
+import com.scg.stop.notice.domain.Notice;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class EventNoticeResponse {
+public class NoticeResponse {
     private Long id;
     private String title;
     private String content;
@@ -23,21 +24,21 @@ public class EventNoticeResponse {
     private LocalDateTime updatedAt;
     private List<FileResponse> files;
 
-    // create a new EventNoticeResponse object
-    public static EventNoticeResponse from(EventNotice eventNotice, List<File> files) {
+    // create a new NoticeResponse object
+    public static NoticeResponse from(Notice notice, List<File> files) {
 
         List<FileResponse> fileResponses = files.stream()
                 .map(FileResponse::from)
-                .collect(java.util.stream.Collectors.toList());
+                .collect(Collectors.toList());
 
-        return new EventNoticeResponse(
-                eventNotice.getId(),
-                eventNotice.getTitle(),
-                eventNotice.getContent(),
-                eventNotice.getHitCount(),
-                eventNotice.isFixed(),
-                eventNotice.getCreatedAt(),
-                eventNotice.getUpdatedAt(),
+        return new NoticeResponse(
+                notice.getId(),
+                notice.getTitle(),
+                notice.getContent(),
+                notice.getHitCount(),
+                notice.isFixed(),
+                notice.getCreatedAt(),
+                notice.getUpdatedAt(),
                 fileResponses
         );
     }
