+++ src/main/java/com/scg/stop/notice/repository/NoticeRepository.java
@@ -1,7 +1,7 @@
-package com.scg.stop.domain.notice.repository;
+package com.scg.stop.notice.repository;
 
-import com.scg.stop.domain.notice.domain.Notice;
-import com.scg.stop.domain.notice.dto.response.NoticeListElementResponse;
+import com.scg.stop.notice.domain.Notice;
+import com.scg.stop.notice.dto.response.NoticeListElementResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -15,7 +15,7 @@ public interface NoticeRepository extends JpaRepository<Notice, Long> {
     @Query("SELECT n FROM Notice n " +
             "WHERE (:title IS NULL OR n.title LIKE %:title%) AND  n.fixed = false")
     Page<NoticeListElementResponse> findNonFixedNotices(@Param("title") String title, Pageable pageable);
-    
+
 
     @Query("SELECT n FROM Notice n " +
             "WHERE (:title IS NULL OR n.title LIKE %:title%) AND n.fixed = true")
