+++ src/test/java/com/scg/stop/domain/notice/controller/NoticeControllerTest.java
@@ -3,10 +3,11 @@ package com.scg.stop.domain.notice.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.scg.stop.configuration.AbstractControllerTest;
 import com.scg.stop.domain.file.dto.response.FileResponse;
-import com.scg.stop.domain.notice.dto.request.NoticeRequest;
-import com.scg.stop.domain.notice.dto.response.NoticeListElementResponse;
-import com.scg.stop.domain.notice.dto.response.NoticeResponse;
-import com.scg.stop.domain.notice.service.NoticeService;
+import com.scg.stop.notice.controller.NoticeController;
+import com.scg.stop.notice.dto.request.NoticeRequest;
+import com.scg.stop.notice.dto.response.NoticeListElementResponse;
+import com.scg.stop.notice.dto.response.NoticeResponse;
+import com.scg.stop.notice.service.NoticeService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
